refactor: move the package [#22]
