+++ spring/src/main/java/com/bancow/bancowback/common/util/mail/repository/MailTemplateRepository.java
@@ -1,11 +1,11 @@
-package com.bancow.bancowback.util.mail.repository;
+package com.bancow.bancowback.common.util.mail.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.bancow.bancowback.util.mail.entity.MailTemplate;
+import com.bancow.bancowback.common.util.mail.entity.MailTemplate;
 
 @Repository
 public interface MailTemplateRepository extends JpaRepository<MailTemplate, Long> {
