+++ spring/src/main/java/com/bancow/bancowback/common/util/token/entity/Token.java
@@ -1,11 +1,16 @@
-package com.bancow.bancowback.util.mail.entity;
+package com.bancow.bancowback.common.util.token.entity;
 
 import java.time.LocalDateTime;
 
+import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.OneToOne;
+
+import com.bancow.bancowback.manager.entity.Manager;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -19,18 +24,17 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-public class MailTemplate {
+public class Token {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	private String templateId;
+	@Column
+	private String token;
 
-	private String title;
-	private String contents;
+	@OneToOne(fetch = FetchType.LAZY)
+	private Manager manager;
 
-	private String sendEmail;
-	private String sendUserName;
-
-	private LocalDateTime regDate;
+	@Column(name = "expired_date")
+	private LocalDateTime expiredDate;
 }
