+++ spring/src/main/java/com/bancow/bancowback/common/util/token/repository/TokenRepository.java
@@ -1,11 +1,11 @@
-package com.bancow.bancowback.util.token.repository;
+package com.bancow.bancowback.common.util.token.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.bancow.bancowback.util.token.entity.Token;
+import com.bancow.bancowback.common.util.token.entity.Token;
 
 @Repository
 public interface TokenRepository extends JpaRepository<Token, Long> {
